+++ core-build/gulp-core-build-webpack/src/index.ts
@@ -3,6 +3,7 @@ import { WebpackTask } from './WebpackTask';
 export { IWebpackTaskConfig, WebpackTask } from './WebpackTask'
 
 /* tslint:disable:export-name */
+/** @public */
 export const webpack: WebpackTask = new WebpackTask();
 export default webpack;
 /* tslint:enable:export-name */