+++ libraries/api-extractor/testInputs/example1/example1-output.api.ts
@@ -1,9 +1,24 @@
+// WARNING: The underscore prefix ("_") should only be used with definitions that are explicitly marked as @internal
+// @public
 class ___proto__ {
   // (undocumented)
   public propertyIsEnumerable: string;
 }
 
-// (undocumented)
+// WARNING: propertyWithNoType has incomplete type information
+// @internal
+class _InternalClass {
+  public _internalMethodWithRedundantUnderscore(): void;
+  // (undocumented)
+  public test(): void;
+}
+
+// @internal (preapproved)
+class _PreapprovedInternalClass {
+}
+
+
+// @public (undocumented)
 class A extends __proto__, implements hasOwnProperty {
   // (undocumented)
   ___lookupSetter__: __proto__;
@@ -13,10 +28,10 @@ class A extends __proto__, implements hasOwnProperty {
 
 // @public
 class AliasClass4 {
+  // @internal
+  public _aliasFunc(): void;
   // (undocumented)
   public aliasField: number;
-  // @internal
-  public aliasFunc(): void;
   // (undocumented)
   public readonly shouldBeReadOnly: number;
 }
@@ -30,27 +45,26 @@ class AlphaTaggedClass {
 
 // @beta
 class BetaTaggedClass {
+  // WARNING: The underscore prefix ("_") should only be used with definitions that are explicitly marked as @internal
+  // @alpha
+  public _alphaMethodWithBadUnderscore(): void;
   // @internal
   public _internalMethod(): void;
   // @alpha
   public alphaMethod(): void;
+  // WARNING: Because this definition is explicitly marked as @internal, an underscore prefix ("_") should be added to its name
+  // @internal
+  public internalMethodMissingUnderscore(): void;
   public plainMethod(): void;
 }
 
-// (undocumented)
+// @public (undocumented)
 interface hasOwnProperty {
   // (undocumented)
   ___lookupSetter__: __proto__;
 }
 
-// WARNING: propertyWithNoType has incomplete type information
-// @internal
-class InternalClass {
-  // (undocumented)
-  public test(): void;
-}
-
-// (undocumented)
+// @public (undocumented)
 class MyClass {
   // (undocumented)
   public field: number;
@@ -60,12 +74,7 @@ class MyClass {
   public test(): void;
 }
 
-// @internal (preapproved)
-class PreapprovedInternalClass {
-}
-
-
-// (undocumented)
+// @public (undocumented)
 export function publicFunction(param: number): string;
 
 // @public