+++ libraries/package-deps-hash/src/getPackageDeps.ts
@@ -1,8 +1,113 @@
 import * as child_process from 'child_process';
 import { IPackageDeps } from './IPackageDeps';
 
-const PROCESS_OUTPUT_DELIMITER: string = '///~X~X~X~X~X~X~///';
+export type GitStatusChangeType = 'D' | 'M' | 'A';
 
+/**
+ * Parses the output of the "git ls-tree" command
+ */
+export function parseGitLsTree(output: string): Map<string, string> {
+  const changes: Map<string, string> = new Map<string, string>();
+
+  if (output) {
+    // Note: The output of git ls-tree uses \n newlines regardless of OS.
+    output.split('\n').forEach(line => {
+
+      // A line is expected to look like:
+      // 100644 blob 3451bccdc831cb43d7a70ed8e628dcf9c7f888c8    src/typings/tsd.d.ts
+
+      if (line) {
+        // Take everything after the "100644 blob", which is just the hash and filename
+        const [hash, filename]: string[] = line.substr(line.indexOf('blob ') + 5).split('\t');
+        if (!filename || !hash) {
+          throw new Error(`Cannot parse git ls-tree input: "${line}"`);
+        }
+
+        changes.set(filename, hash);
+      }
+    });
+  }
+
+  return changes;
+}
+
+/**
+ * Parses the output of the "git status" command
+ */
+export function parseGitStatus(output: string, packagePath: string): Map<string, GitStatusChangeType> {
+  const changes: Map<string, GitStatusChangeType> = new Map<string, GitStatusChangeType>();
+  const filesToHash: string[] = [];
+
+  /*
+  * Typically, output will look something like:
+  * M temp_modules/rush-package-deps-hash/package.json
+  * D package-deps-hash/src/index.ts
+  */
+
+  // If there was an issue with `git ls-tree`, or there are no current changes, processOutputBlocks[1]
+  // will be empty or undefined
+  if (!output) {
+    return changes;
+  }
+
+  // Note: The output of git hash-object uses \n newlines regardless of OS.
+  output
+    .trim()
+    .split('\n')
+    .forEach(line => {
+      const [changeType, filename]: string[] = line.trim().split(' ');
+      /*
+      * changeType == 'D' or 'M' or 'A'
+      * filename == path to the file
+      */
+      if (changeType && filename) {
+        changes.set(filename, changeType as GitStatusChangeType);
+      }
+    });
+
+  return changes;
+}
+
+/**
+ * Takes a list of files and returns the current git hashes for them
+ */
+export function gitHashFiles(filesToHash: string[], packagePath: string): Map<string, string> {
+  const changes: Map<string, string> = new Map<string, string>();
+  if (filesToHash.length) {
+    const hashStdout: string = child_process.execSync(
+      'git hash-object ' + filesToHash.join(' '),
+      { cwd: packagePath }).toString();
+
+    // The result of hashStdout will be a list of file hashes delimited by newlines
+    const hashes: string[] = hashStdout.split('\n');
+
+    filesToHash.forEach((filename, i) => changes.set(filename, hashes[i]));
+  }
+  return changes;
+}
+
+/**
+ * Executes "git ls-tree" in a folder
+ */
+export function gitLsTree(path: string): string {
+  return child_process.execSync(
+    `git ls-tree HEAD -r`,
+    { cwd: path }).toString();
+}
+
+/**
+ * Executes "git status" in a folder
+ */
+export function gitStatus(path: string): string {
+  return child_process.execSync(
+    `git status -s -u .`,
+    { cwd: path }).toString();
+}
+
+/**
+ * Collects the current git filehashes for a directory
+ * @public
+ */
 export function getPackageDeps(packagePath: string = process.cwd(), excludedPaths?: string[]): IPackageDeps {
   const excludedHashes: { [key: string]: boolean } = {};
 
@@ -10,76 +115,38 @@ export function getPackageDeps(packagePath: string = process.cwd(), excludedPath
     excludedPaths.forEach(path => excludedHashes[path] = true);
   }
 
-  const stdout: string = child_process.execSync(
-    `git ls-tree HEAD -r && echo ${PROCESS_OUTPUT_DELIMITER} && git status -s -u .`,
-    { cwd: packagePath }).toString();
-
   const changes: IPackageDeps = {
     files: {}
   };
-  const processOutputBlocks: string[] = stdout.split(PROCESS_OUTPUT_DELIMITER);
 
-  // Note: The output of git ls-tree uses \n newlines regardless of OS.
-  processOutputBlocks[0].split('\n').forEach(line => {
+  const gitLsOutput: string = gitLsTree(packagePath);
 
-    // A line is expected to look like:
-    // 100644 blob 3451bccdc831cb43d7a70ed8e628dcf9c7f888c8    src/typings/tsd.d.ts
+  // Add all the checked in hashes
+  parseGitLsTree(gitLsOutput).forEach((hash: string, filename: string) => {
+    if (!excludedHashes[filename]) {
+      changes.files[filename] = hash;
+    }
+  });
 
-    if (line) {
-      // Take everything after the "100644 blob", which is just the hash and filename
-      const [hash, filename]: string[] = line.substr(line.indexOf('blob ') + 5).split('\t');
+  // Update the checked in hashes with the current repo status
+  const gitStatusOutput: string = gitStatus(packagePath);
+  const currentlyChangedFiles: Map<string, GitStatusChangeType > =
+    parseGitStatus(gitStatusOutput, packagePath);
+
+  const filesToHash: string[] = [];
+  currentlyChangedFiles.forEach((changeType: GitStatusChangeType, filename: string) => {
+    if (changeType === 'D') {
+      delete changes.files[filename];
+    } else {
       if (!excludedHashes[filename]) {
-        changes.files[filename] = hash;
+        filesToHash.push(filename);
       }
     }
   });
 
-  // If there was an issue with `git ls-tree`, or there are no current changes, processOutputBlocks[1]
-  // will be empty or undefined
-  if (processOutputBlocks[1]) {
-    const filesToHash: string[] = [];
+  gitHashFiles(filesToHash, packagePath).forEach((hash: string, filename: string) => {
+    changes.files[filename] = hash;
+  });
 
-    /*
-     * Typically, processOutputBlocks[1] will look something like:
-     * M temp_modules/rush-package-deps-hash/package.json
-     * D package-deps-hash/src/index.ts
-     */
-
-    // Note: The output of git hash-object uses \n newlines regardless of OS.
-    processOutputBlocks[1]
-      .trim()
-      .split('\n')
-      .forEach(line => {
-        const [changeType, filename]: string[] = line.trim().split(' ');
-        /*
-         * changeType == 'D' or 'M' or 'A'
-         * filename == path to the file
-         */
-
-        if (changeType && filename) {
-          // If the file is currently deleted, then it will have a 'D'
-          if (changeType === 'D') {
-            delete changes.files[filename];
-          } else {
-            // Otherwise the file was changed or added and we should get the current hash
-            if (!excludedHashes[filename]) {
-              filesToHash.push(filename);
-            }
-          }
-        }
-      });
-
-    if (filesToHash.length) {
-      const hashStdout: string = child_process.execSync(
-        'git hash-object ' + filesToHash.join(' '),
-        { cwd: packagePath }).toString();
-
-      // The result of hashStdout will be a list of file hashes delimited by newlines
-
-      const hashes: string[] = hashStdout.split('\n');
-
-      filesToHash.forEach((filename, i) => changes.files[filename] = hashes[i]);
-    }
-  }
   return changes;
 }